Extract waveform selection paint helpers
